+++ lib/decorators/routes.ts
@@ -1,5 +1,0 @@
-import { Route } from '../service'
-
-export const Routes = (routes: Route[]) => (constructor: Function): void => {
-  constructor.prototype.definedRoutes = routes
-}