+++ tests/before.spec.ts
@@ -1,17 +1,14 @@
-import chai from 'chai'
 import { Request, Response, Service } from '../lib'
 import { Before } from '../lib'
 import buildMockRequest from './utils/buildMockRequest'
 
-const expect = chai.expect
-
 describe('@Before decorator', () => {
-  it('should correctly pass in the request', async () => {
-    let reqUserId: string
+  let reqUserId: string
 
+  it('should correctly pass in the request', async () => {
     class UserService extends Service {
       @Before(async (req: Request): Promise<void> => {
-        reqUserId = req.query.userId
+        reqUserId = req.query?.userId ?? ''
       })
       async get(req: Request): Promise<Response> {
         return {
@@ -30,8 +27,6 @@ describe('@Before decorator', () => {
   })
 
   it('should correctly pass the caller context', async () => {
-    let reqUserId: string
-
     class UserService extends Service {
       @Before(async (req: Request, caller: UserService): Promise<void> => {
         caller.handleBefore(req)
@@ -43,7 +38,7 @@ describe('@Before decorator', () => {
       }
 
       handleBefore(req: Request): void {
-        reqUserId = req.query.userId
+        reqUserId = req.query?.userId ?? ''
       }
     }
 
@@ -60,7 +55,7 @@ describe('@Before decorator', () => {
     class UserService extends Service {
       @Before(async (req: Request): Promise<void> => {
         req.ctx.state.user = {
-          id: Number(req.query.userId)
+          id: Number(req.query?.userId ?? '')
         }
       })
       async get(req: Request): Promise<Response> {