+++ tests/required-hook.spec.ts
@@ -1,14 +1,11 @@
-import chai from 'chai'
 import { Service, Request, Response, Validate, Required } from '../lib'
 import buildMockRequest from './utils/buildMockRequest'
-
-const expect = chai.expect
 
 describe('@Required decorator', () => {
   it('should return a single error for a missing property with no configuration on a post request', async () => {
     class Stat {
       @Required()
-      name: string
+      name: string = ''
     }
 
     class StatService extends Service {
@@ -29,14 +26,14 @@ describe('@Required decorator', () => {
 
     expect(res.status).to.equal(400)
 
-    expect(res.body.errors).to.have.key('name')
-    expect(res.body.errors.name).to.eql(['name is missing from the request body'])
+    expect(res.body?.errors).to.have.key('name')
+    expect(res.body?.errors.name).to.eql(['name is missing from the request body'])
   })
 
   it('should not return an error if the method does not match the default methods post and put', async () => {
     class Stat {
       @Required()
-      name: string
+      name: string = ''
     }
 
     class StatService extends Service {
@@ -61,7 +58,7 @@ describe('@Required decorator', () => {
   it('should not return an error if the required fields are present', async () => {
     class Stat {
       @Required()
-      name: string
+      name: string = ''
     }
 
     class StatService extends Service {
@@ -89,7 +86,7 @@ describe('@Required decorator', () => {
   it('should not return an error if a required field is null', async () => {
     class Stat {
       @Required()
-      name: string
+      name: string = ''
     }
 
     class StatService extends Service {
@@ -117,7 +114,7 @@ describe('@Required decorator', () => {
   it('should return an error if a field alias is not present', async () => {
     class Stat {
       @Required({ as: 'entityName' })
-      name: string
+      name: string = ''
     }
 
     class StatService extends Service {
@@ -141,14 +138,14 @@ describe('@Required decorator', () => {
 
     expect(res.status).to.equal(400)
 
-    expect(res.body.errors).to.have.key('entityName')
-    expect(res.body.errors.entityName).to.eql(['entityName is missing from the request body'])
+    expect(res.body?.errors).to.have.key('entityName')
+    expect(res.body?.errors.entityName).to.eql(['entityName is missing from the request body'])
   })
 
   it('should not return an error if a field alias is present', async () => {
     class Stat {
       @Required({ as: 'entityName' })
-      name: string
+      name: string = ''
     }
 
     class StatService extends Service {
@@ -175,7 +172,7 @@ describe('@Required decorator', () => {
 
   it('should not return an error if there are no required properties', async () => {
     class Stat {
-      name: string
+      name: string = ''
     }
 
     class StatService extends Service {
@@ -200,12 +197,12 @@ describe('@Required decorator', () => {
 
   it('should let requiredIf take precedent over the request method requirement', async () => {
     class User {
-      firstName: string
+      firstName: string = ''
 
       @Required({
         requiredIf: (req: Request) => req.body.firstName
       })
-      lastName: string
+      lastName: string = ''
     }
 
     class UserService extends Service {